feat: add assertions for email type
